WIP: Start moving some code out of that monolithic VC
