+++ HowManyCities/Stats/ViewModels/GameStatsViewModel.swift
@@ -1,0 +1,12 @@
+//
+//  GameStatsViewModel.swift
+//  HowManyCities
+//
+//  Created by Geoffrey Liu on 5/10/22.
+//
+
+import Foundation
+
+final class GameStatsViewModel {
+  
+}
